backblaze_requirements.txt
django-storages==1.13.2
boto3==1.28.38
python-dotenv==1.0.0
